Ìnitial commit of scripts
These Python scripts are not new, but it's time to put them under
version control. They convert data from the Belgian public transport
companies De Lijn and TEC to a format suitable for uploading to
Openstreetmap.
Recently the scripts were modified to use ref:De_Lijn, ref:TECB etc
instead of simply ref, route_ref and zone, in order to make it possible
to have 1 node serve for more than 1 operator.
Another optimisation was the use of separate tables for data of De Lijn,
TEC and OSM. meaning the tables can simply be dropped and the new data
COPY'd in instead of performing updates.